accounting/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.6
django-cors-headers==4.4.0
django-crispy-forms==2.2
django-scheduler==0.10.1
django-watchman==1.3.0
djangorestframework==3.15.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.0
icalendar==5.0.13
idna==3.7
multidict==6.0.5
numpy==1.26.4
openai==0.28.0
packaging==24.1
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rsa==4.9
safetensors==0.4.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
